Separate rebuildDb function from running it with seed.js
